# swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "Riddle of Sphinx 2021 API",
      "version": "1.0.0",
      "title": "K! ROS"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
    "/user/resetPassword": {
        "post": {
            "tags": [
                "user"
            ],
            "summary": "Resets the password",
            "consumes": [
                "application/json"
            ],
            "produces": [
                "application/json"
            ],
            "parameters": [
                {
                    "in": "body",
                    "name": "body",
                    "description": "Resets the password",
                    "required": true,
                    "schema": {
                    "$ref": "#/definitions/ResetPassword"
                    }
                }
            ],
            "responses": {
                "200": {
                    "description": "Successful Login"
                },
                "400": {
                    "description": "Server error"
                },
                "401": {
                    "description": "User(email) not found"
                }
            }
        }
    }
    },
    "definitions": {
...
    "ResetPassword": {
        "type": "object",
        "properties": {
            "password": {
                "type": "string"
            }
        }
      }
    },
    "securityDefinitions": {
        "auth" : {
            "type": "apiKey",
            "name": "authorization",
            "in": "header"
          }
      }
  }